refactor: 비즈니스 로직이 들어간 유효성 검사를 서비스 계층에 위치
- 커스텀 Validator에서 repository에 의존하여 비즈니스 로직을 수행하는 것은 적절치 않다고 판단
