+++ src/main/java/com/peeerr/climbing/dto/request/CategoryCreateRequest.java
@@ -1,6 +1,5 @@
 package com.peeerr.climbing.dto.request;
 
-import com.peeerr.climbing.validation.annotation.NotDuplicateCategoryName;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Size;
 import lombok.AccessLevel;
@@ -8,7 +7,6 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-import static com.peeerr.climbing.exception.ErrorMessage.ALREADY_EXISTS_CATEGORY;
 import static com.peeerr.climbing.exception.ErrorMessage.CATEGORY_NAME_NOT_BLANK;
 
 @NoArgsConstructor
@@ -18,7 +16,6 @@ public class CategoryCreateRequest {
 
     @Size(min = 1, max = 20)
     @NotBlank(message = CATEGORY_NAME_NOT_BLANK)
-    @NotDuplicateCategoryName(message = ALREADY_EXISTS_CATEGORY)
     private String categoryName;
 
     public static CategoryCreateRequest of(String categoryName) {
